unsupervised/requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9.0
datasets
Pillow
numpy
matplotlib
scikit-learn
PyYAML
tqdm
